cleanup dot/sync
