+++ dot/sync/syncer.go
@@ -30,12 +30,6 @@ import (
 	log "github.com/ChainSafe/log15"
 )
 
-var (
-	maxInt64                     = int64(2 ^ 63 - 1)
-	blockRequestSize      uint32 = 128
-	blockRequestQueueSize int64  = 3
-)
-
 // Service deals with chain syncing by sending block request messages and watching for responses.
 type Service struct {
 	logger log.Logger
@@ -111,31 +105,6 @@ func NewService(cfg *Config) (*Service, error) {
 	}, nil
 }
 
-// // HandleBlockAnnounceHandshake handles a block that a peer claims to have through a HandleBlockAnnounceHandshake
-// func (s *Service) HandleBlockAnnounceHandshake(blockNum *big.Int) []*network.BlockRequestMessage {
-// 	bestNum, err := s.blockState.BestBlockNumber()
-// 	if err != nil {
-// 		s.logger.Error("failed to get best block number", "error", err)
-// 		return nil // TODO: handle this / panic?
-// 	}
-
-// 	if blockNum == nil || bestNum.Cmp(blockNum) != -1 || s.highestSeenBlock.Cmp(blockNum) >= 0 {
-// 		return nil
-// 	}
-
-// 	s.highestSeenBlock = blockNum
-
-// 	// need to sync
-// 	start := new(big.Int).Add(bestNum, big.NewInt(1)).Int64()
-
-// 	if s.synced {
-// 		s.synced = false
-// 		_ = s.blockProducer.Pause()
-// 	}
-
-// 	return s.createBlockRequests(start, s.highestSeenBlock.Int64())
-// }
-
 // HandleBlockAnnounce creates a block request message from the block
 // announce messages (block announce messages include the header but the full
 // block is required to execute `core_execute_block`).
@@ -173,143 +142,10 @@ func (s *Service) HandleBlockAnnounce(msg *network.BlockAnnounceMessage) error {
 		)
 	}
 
-	// // check if block body is stored in block state (ie. if we have the full block already)
-	// has, _ = s.blockState.HasBlockBody(header.Hash())
-	// if !has {
-	// 	s.synced = false
-	// 	err = s.blockProducer.Pause()
-	// 	if err != nil {
-	// 		s.logger.Debug("failed to pause block production")
-	// 	}
-
-	// 	// create block request to send
-	// 	bestNum, err := s.blockState.BestBlockNumber() //nolint
-	// 	if err != nil {
-	// 		s.logger.Error("failed to get best block number", "error", err)
-	// 		bestNum = big.NewInt(0)
-	// 	}
-
-	// 	// if we already have blocks up to the BlockAnnounce number, only request the block in the BlockAnnounce
-	// 	var start int64
-	// 	if bestNum.Cmp(header.Number) > 0 {
-	// 		start = header.Number.Int64()
-	// 	} else {
-	// 		s.highestSeenBlock = header.Number
-	// 		start = bestNum.Int64() + 1
-	// 	}
-
-	// 	return s.createBlockRequest(start, 1)
-	// } else if err != nil {
-	// 	s.logger.Error("failed to handle BlockAnnounce", "error", err)
-	// }
-
 	return nil
 }
 
-// HandleBlockResponse handles a BlockResponseMessage by processing the blocks found in it and adding them to the BlockState if necessary.
-// If the node is still not synced after processing, it creates and returns the next BlockRequestMessage to send.
-func (s *Service) HandleBlockResponse(msg *network.BlockResponseMessage) error {
-	// highestInResp will be the highest block in the response
-	// it's set to 0 if err != nil
-	//for _, bd := range msg.BlockData {
-	err := s.ProcessBlockData(msg.BlockData)
-	if err != nil {
-		return err
-	}
-	//}
-
-	return nil
-	// var start int64
-	// low, high, err := s.processBlockResponseData(msg)
-
-	// // if we cannot find the parent block in our blocktree, we are missing some blocks, and need to request
-	// // blocks from farther back in the chain
-	// if err == blocktree.ErrParentNotFound || errors.Is(err, chaindb.ErrKeyNotFound) {
-	// 	s.logger.Debug("got ErrParentNotFound or ErrKeyNotFound; need to request earlier blocks")
-	// 	bestNum, err := s.blockState.BestBlockNumber() //nolint
-	// 	if err != nil {
-	// 		s.logger.Error("failed to get best block number", "error", err)
-	// 		start = low - maxResponseSize
-	// 	} else {
-	// 		start = bestNum.Int64() + 1
-	// 	}
-
-	// 	s.logger.Debug("retrying block request", "start", start)
-	// 	return s.createBlockRequest(start, blockRequestSize)
-	// } else if err != nil {
-	// 	s.logger.Error("failed to process block response", "error", err)
-	// 	return nil
-	// }
-
-	// s.logger.Debug("received BlockResponse", "start", low, "end", high)
-
-	// // TODO: max retries before unlocking BlockProducer, in case no response is received
-	// bestNum, err := s.blockState.BestBlockNumber()
-	// if err != nil {
-	// 	s.logger.Error("failed to get best block number", "error", err)
-	// 	bestNum = big.NewInt(0)
-	// }
-
-	// // check if we are synced or not
-	// if bestNum.Cmp(s.highestSeenBlock) >= 0 && bestNum.Cmp(big.NewInt(0)) != 0 {
-	// 	s.logger.Debug("all synced up!", "number", bestNum)
-	// 	s.benchmarker.end(uint64(bestNum.Int64()))
-
-	// 	if !s.synced {
-	// 		err = s.blockProducer.Resume()
-	// 		if err != nil {
-	// 			s.logger.Warn("failed to resume block production")
-	// 		}
-	// 		s.synced = true
-	// 	}
-	// 	return nil
-	// }
-
-	// // not yet synced, send another block request for the following blocks
-	// start = bestNum.Int64() + 1
-	// return s.createBlockRequest(start, blockRequestSize)
-}
-
-// func (s *Service) createBlockRequests(start, end int64) []*network.BlockRequestMessage {
-// 	numReqs := (end - start) / int64(blockRequestSize)
-// 	if numReqs > blockRequestQueueSize {
-// 		numReqs = blockRequestQueueSize
-// 	}
-
-// 	if end-start < int64(blockRequestSize) {
-// 		numReqs = 1
-// 	}
-
-// 	reqs := make([]*network.BlockRequestMessage, numReqs)
-// 	for i := 0; i < int(numReqs); i++ {
-// 		offset := i * int(blockRequestSize)
-// 		reqs[i] = s.createBlockRequest(start+int64(offset), blockRequestSize)
-// 	}
-// 	return reqs
-// }
-
-// func (s *Service) createBlockRequest(startInt int64, size uint32) *network.BlockRequestMessage {
-// 	start, err := variadic.NewUint64OrHash(uint64(startInt))
-// 	if err != nil {
-// 		s.logger.Error("failed to create block request start block", "error", err)
-// 		return nil
-// 	}
-
-// 	s.logger.Debug("creating block request", "start", start)
-
-// 	blockRequest := &network.BlockRequestMessage{
-// 		RequestedData: network.RequestedDataHeader + network.RequestedDataBody + network.RequestedDataJustification,
-// 		StartingBlock: start,
-// 		EndBlockHash:  optional.NewHash(false, common.Hash{}),
-// 		Direction:     0, // ascending
-// 		Max:           optional.NewUint32(true, size),
-// 	}
-
-// 	s.benchmarker.begin(uint64(startInt))
-// 	return blockRequest
-// }
-
-// ProcessBlockData processes the BlockResponse and returns the index of the last BlockData it successfully handled.
+// ProcessBlockData processes the BlockData from a BlockResponse and returns the index of the last BlockData it successfully handled.
 func (s *Service) ProcessBlockData(data []*types.BlockData) error {
 	if len(data) == 0 {
 		return ErrNilBlockData
@@ -345,14 +181,6 @@ func (s *Service) ProcessBlockData(data []*types.BlockData) error {
 			}
 
 			s.logger.Trace("header processed", "hash", bd.Hash)
-
-			// if header.Number.Int64() < start {
-			// 	start = header.Number.Int64()
-			// }
-
-			// if header.Number.Int64() > end {
-			// 	end = header.Number.Int64()
-			// }
 		}
 
 		if bd.Body.Exists() && !hasBody {
